+++ tests/universal/Recipe/Step/EndLoopingTest.php
@@ -1,0 +1,80 @@
+<?php
+
+/*
+ * East Common.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license
+ * it is available in LICENSE file at the root of this package
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richard@teknoo.software so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ *
+ * @link        http://teknoo.software/east/common Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+  */
+
+declare(strict_types=1);
+
+namespace Teknoo\Tests\East\Common\Recipe\Step;
+
+use PHPUnit\Framework\TestCase;
+use Teknoo\East\Common\Recipe\Step\EndLooping;
+use Teknoo\East\Common\Recipe\Step\StartLoopingOn;
+use Teknoo\East\Foundation\Manager\ManagerInterface;
+
+/**
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ * @covers \Teknoo\East\Common\Recipe\Step\EndLooping
+ */
+class EndLoopingTest extends TestCase
+{
+    public function buildStep(): EndLooping
+    {
+        return new EndLooping();
+    }
+
+    public function testInvokeLoopNotEnded()
+    {
+        $manager = $this->createMock(ManagerInterface::class);
+        $manager->expects(self::once())
+            ->method('continue')
+            ->with([], StartLoopingOn::class)
+            ->willReturnSelf();
+
+        $loop = $this->createMock(StartLoopingOn::class);
+        $loop->expects(self::any())
+            ->method('isEnded')
+            ->willReturn(false);
+
+        self::assertInstanceOf(
+            EndLooping::class,
+            $this->buildStep()($manager, $loop)
+        );
+    }
+
+    public function testInvokeLoopEnded()
+    {
+        $manager = $this->createMock(ManagerInterface::class);
+        $manager->expects(self::never())
+            ->method('continue');
+
+        $loop = $this->createMock(StartLoopingOn::class);
+        $loop->expects(self::any())
+            ->method('isEnded')
+            ->willReturn(true);
+
+        self::assertInstanceOf(
+            EndLooping::class,
+            $this->buildStep()($manager, $loop)
+        );
+    }
+}